create separate impls for syscall4
